Use `Interceptor.attach` for `applierWithParameters`
Before we used Interceptor.replace which called `NativeFunction`.
For reasons that are not very clear to me, it could throw "abort was called", which made us return a null pointer and possibly crash the program.
Now we use Interceptor.attach and thus do not interrupt the original execution.

For example:

previous implementation
il2cpp:
0x010d4c94 ┌─UserTargeting.HamsterOfferwall::get_ActiveReward(this = HamsterOfferwall (UserTargeting.HamsterOfferwall))
0x010d4c94 └─UserTargeting.HamsterOfferwall::get_ActiveReward = null [native IL2CPP exception occurred] // <- null pointer returned

after that the app crashed.

new implementation
il2cpp:
0x010d4c94 ┌─UserTargeting.HamsterOfferwall::get_ActiveReward(this = HamsterOfferwall (UserTargeting.HamsterOfferwall))
0x010d4c94 └─UserTargeting.HamsterOfferwall::get_ActiveReward = ""

and execution continues as desired, yay!

One more important improvement, as @fakekey and @Vladik01-11 noticed, some apps have issues with getting the parameter value (again, for reasons I don't understand).
Now we check if it is possible to get the parameter value, instead of spamming the console with errors, if it is impossible to get the parameter value (as I noticed, this applies mostly to Il2Cpp.Object) you will see the red text "null (failed to parse parameter value)".